Move output to sql to function
moved output to sql to seperate orchestrator
