+++ shared_code/__init__.py
@@ -1,3 +1,3 @@
 """Initialize package"""
 
-from shared_code import get_config
+from shared_code import get_config, list_to_string
